+++ packages/prettier/src/types.ts
@@ -1,0 +1,7 @@
+import type Prettier from 'prettier';
+
+export interface PrettierOptions extends PrettierConfig {
+  tailwindcss: boolean;
+}
+
+export interface PrettierConfig extends Prettier.Config {}