+++ prettier.config.js
@@ -1,5 +1,3 @@
 import prettierConfig from '@jsxiaosi/eslint-config-prettier';
 
-export default {
-  ...prettierConfig,
-};
+export default prettierConfig();